graphics: align remaining WORD/LONG LVO parameter mismatches with graphics.conf
Follow-up to PR #822 (ScrollRaster/ScrollRasterBF negative-delta fix). A
scripted sweep of every graphics.conf functionlist entry against its AROS_LH*
implementation found the same narrow(conf)/wide(impl) parameter-type mismatch
in eleven more graphics.library functions: BltBitMap, EraseRect, RectFill,
Flood, ReadPixelLine8, WritePixelLine8, WriteChunkyPixels, WritePixelArray8,
and the private WritePixels8, FillRectPenDrMd and DoPixelFunc.

graphics.conf declares these coordinates WORD, so every generated caller stub
passes 16-bit arguments; the implementations read LONG or ULONG. Under AAPCS64
(and any ABI where the caller is not required to sign-extend narrow arguments)
a negative coordinate can arrive as a large positive value in the callee,
exactly the ScrollRaster failure mode. Make the implementation parameter types
match the .conf so the callee sign-extends the low 16 bits itself; already-built
caller binaries stay correct.

The sweep also covered workbench/libs and compiler; graphics.library was the
only remaining offender.

+++ rom/graphics/flood.c
@@ -47,8 +47,8 @@ static int pix_written;
 /*  SYNOPSIS */
         AROS_LHA(struct RastPort *, rp, A1),
         AROS_LHA(ULONG            , mode, D2),
-        AROS_LHA(LONG             , x, D0),
-        AROS_LHA(LONG             , y, D1),
+        AROS_LHA(WORD             , x, D0),
+        AROS_LHA(WORD             , y, D1),
 
 /*  LOCATION */
         struct GfxBase *, GfxBase, 55, Graphics)
