+++ rom/graphics/writepixelline8.c
@@ -14,9 +14,9 @@
 
 /*  SYNOPSIS */
         AROS_LHA(struct RastPort *      , rp            , A0),
-        AROS_LHA(LONG                   , xstart        , D0),
-        AROS_LHA(LONG                   , ystart        , D1),
-        AROS_LHA(ULONG                  , width         , D2),
+        AROS_LHA(WORD                   , xstart        , D0),
+        AROS_LHA(WORD                   , ystart        , D1),
+        AROS_LHA(WORD                   , width         , D2),
         AROS_LHA(UBYTE *                , array         , A2),
         AROS_LHA(struct RastPort *      , tempRP        , A1),
